pyproject.toml
[tool.poetry]
name = "animegan"
version = "0.1.0"
description = ""
authors = ["Maxim <max.sherb2007@gmail.com>", "Vlad <ogloblinvlad6@gmail.com>", "Andrey <Andrey7Klimin@yandex.ru>", "Andrey <paiflare@gmail.com>", "Yaroslav <blanla12345@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.12.8"
opencv-python = "^4.10.0.84"
aiogram = "^3.15.0"
pillow = "^11.0.0"
python-dotenv = "^1.0.1"
tqdm = "^4.67.1"
sqlalchemy = "^2.0.36"
psycopg2-binary = "^2.9.10"
boto3 = "^1.35.83"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.5"
matplotlib = "^3.10.0"
notebook = "^7.3.1"
torch = "^2.5.1"
torchvision = "^0.20.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
